+++ src/ckanapi_harvesters/__init__.py
@@ -43,7 +43,7 @@ from . import reports
 
 # usage shortcuts
 from .auxiliary import (CkanMap, RequestType, CkanField, CkanState, CkanProgressCallback, CkanCallbackLevel,
-                        CkanProgressBarType, CkanProgressCallbackEmpty)
+                        CkanProgressBarType, CkanProgressCallbackEmpty, CkanCapacity)
 from .policies import PackagePolicyReport, CkanPackageDataFormatPolicy
 from .reports import CkanAdminReport
 from .ckan_api import CkanApi, CkanApiParams, CKAN_API_VERSION